+++ lib/src/session.dart
@@ -19,6 +19,8 @@
 import 'dart:async';
 import 'dart:ffi';
 
+import 'package:cancellation_token/cancellation_token.dart';
+
 import '../realm.dart';
 import 'native/realm_core.dart';
 import 'user.dart';